refactor(INJI-245): move VcItemTags to VC/common folder
