+++ components/VC/ExistingMosipVCItem/ExistingMosipVCItemContent.tsx
@@ -3,7 +3,7 @@ import {useTranslation} from 'react-i18next';
 import {Image, ImageBackground, View} from 'react-native';
 import {getLocalizedField} from '../../../i18n';
 import {VerifiableCredential} from '../../../types/VC/ExistingMosipVC/vc';
-import {VcItemTags} from '../../VcItemTags';
+import {VcItemTags} from '../common/VcItemTags';
 import VerifiedIcon from '../../VerifiedIcon';
 import {Column, Row, Text} from '../../ui';
 import {Theme} from '../../ui/styleUtils';
