+++ components/VC/ExistingMosipVCItem/ExistingMosipVCItemDetails.tsx
@@ -8,7 +8,7 @@ import {VC, CredentialSubject} from '../../../types/VC/ExistingMosipVC/vc';
 import {Button, Column, Row, Text} from '../../ui';
 import {Theme} from '../../ui/styleUtils';
 import {TextItem} from '../../ui/TextItem';
-import {VcItemTags} from '../../VcItemTags';
+import {VcItemTags} from '../common/VcItemTags';
 import VerifiedIcon from '../../VerifiedIcon';
 import {getLocalizedField} from '../../../i18n';
 import {CREDENTIAL_REGISTRY_EDIT} from 'react-native-dotenv';
